+++ app/src/main/AndroidManifest.xml
@@ -66,7 +66,7 @@
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
         <activity
-            android:name=".activity.UpdateCommentActivity"
+            android:name=".activity.UpdateReplyActivity"
             android:theme="@style/AppTheme.NoActionBar">
         </activity>
         <activity
