+++ app/src/main/res/layout/activity_update_reply.xml
@@ -5,7 +5,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".activity.UpdateCommentActivity">
+        tools:context=".activity.UpdateReplyActivity">
 
     <com.google.android.material.appbar.AppBarLayout
             android:layout_height="wrap_content"
@@ -17,7 +17,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="?attr/actionBarSize"
                 android:background="?attr/colorPrimary"
-                app:popupTheme="@style/AppTheme.PopupOverlay"/>
+                app:popupTheme="@style/AppTheme.PopupOverlay" />
 
     </com.google.android.material.appbar.AppBarLayout>
 
@@ -31,7 +31,7 @@
                 android:id="@+id/rv_write"
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:divider="@null"/>
+                android:divider="@null" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
